Flight Info make int field to double.

+++ src/main/java/com/example/atc/data/FlightInfo.java
@@ -5,9 +5,9 @@ import lombok.Data;
 @Data
 public class FlightInfo {
     private String flightName;
-    private int top;
-    private int left;
-    private int right;
-    private int down;
-    private int speed;
+    private double top;
+    private double left;
+    private double right;
+    private double down;
+    private double speed;
 }
